automatic PAGE TURNER
This project details the design and build process of page turning device for people with disability to increase independent. The primary goal of this project is to design and manufacture a product that not only has an efficient operation, but is easy to use for people with disabilities and minimal assistance. The device will consist of a mechanical structure to support the book, an electrical system to facilitate the turning of pages in both directions, and additional features to support comfort and readability, such as Voice control and open the book yourself.